update edit category function, working properly

+++ Feast_Food/src/components/CategoryData.tsx
@@ -3,6 +3,7 @@ import {MdDelete} from "react-icons/md";
 import {useQuery, useMutation} from "@tanstack/react-query";
 import axios from "axios";
 import {useNavigate} from "react-router-dom";
+import { toast } from "react-toastify";
 
 interface CategoryDataProps {
     search: string;
@@ -46,9 +47,13 @@ const CategoryData: React.FC<CategoryDataProps> = ({ search}) => {
                             <td>{i?.name}</td>
                             <td><button className={"edit-btn2"} onClick={()=>{
                                 navigate("/edit/"+i?.id);
+                                console.log(i?.id)
                             }}><CiEdit /></button></td>
-                            <td><button className={"delete-btn2"} onClick={()=>{
-                                deleteByIdApi.mutate(i?.id);
+                            <td><button className={"delete-btn2"} onClick={() => {
+                                // Display confirmation prompt before deletion
+                                if (window.confirm("Are you sure you want to delete this category?")) {
+                                    deleteByIdApi.mutate(i?.id);
+                                }
                             }}><MdDelete /></button></td>
                         </tr>
                     )
